+++ src/Service/SymfonyService.php
@@ -33,10 +33,10 @@ class SymfonyService extends PHPService
         yield [
             'task' => new AsTask('cache-clear', $this->name, 'Clears the application cache'),
             'function' => function (bool $warm = true): void {
-                $this->runInBuilder('rm -rf var/cache');
+                $this->runInBuilder(['rm', '-rf', 'var/cache']);
 
                 if ($warm) {
-                    $this->runInBuilder('php bin/console cache:warm');
+                    $this->runInBuilder(['php', 'bin/console', 'cache:warm']);
                 }
             },
         ];
@@ -44,22 +44,22 @@ class SymfonyService extends PHPService
         yield [
             'task' => new AsTask('cache-warmup', $this->name, 'Warms the application cache'),
             'function' => function (): void {
-                $this->runInBuilder('php bin/console cache:warm');
+                $this->runInBuilder(['php', 'bin/console', 'cache:warm']);
             },
         ];
 
         yield [
             'task' => new AsTask('fixtures', $this->name . ':db', 'Loads fixtures'),
             'function' => function (): void {
-                $this->runInBuilder('php bin/console doctrine:fixture:load');
+                $this->runInBuilder(['php', 'bin/console', 'doctrine:fixture:load']);
             },
         ];
 
         yield [
             'task' => new AsTask('migrate', $this->name . ':db', 'Migrates database schema'),
             'function' => function (): void {
-                $this->runInBuilder('php bin/console doctrine:database:create --if-not-exists');
-                $this->runInBuilder('php bin/console doctrine:migration:migrate -n --allow-no-migration --all-or-nothing');
+                $this->runInBuilder(['php', 'bin/console', 'doctrine:database:create', '--if-not-exists']);
+                $this->runInBuilder(['php', 'bin/console', 'doctrine:migration:migrate', '-n', '--allow-no-migration', '--all-or-nothing']);
             },
         ];
 
@@ -67,7 +67,7 @@ class SymfonyService extends PHPService
             'task' => new AsTask('symfony', $this->name, 'Run a Symfony console command'),
             'function' => function (#[AsRawTokens] array $args): void {
                 docker_exit_code(
-                    'php bin/console ' . implode(' ', array_map(fn($val) => '"' . $val . '"', $args)),
+                    ['php', 'bin/console', ...$args],
                     $this->getBuilderServiceName(),
                     workDir: $this->getBuilderWorkingDirectory(),
                 );